requirements.txt
tqdm
torch
torchtext
tensorflow>=2.0.0-beta0
numpy>=1.16
matplotlib>=3.0.3
sklearn
nltk>=3.3
seaborn>=0.8.1
future
spacy
gpustat
pyyaml
easydict
termcolor
boto3
requests
regex